Site/requirements.txt
absl-py==1.4.0
adafruit-ampy==1.1.0
aiohttp==3.8.4
aioserial==1.3.1
aiosignal==1.3.1
altgraph==0.17.2
appdirs==1.4.4
argcomplete==1.12.3
array-record==0.2.0
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
asyncio==3.4.3
asyncssh==2.14.1
attrs==23.1.0
audioread==3.0.0
bcrypt==4.1.1
beautifulsoup4==4.11.2
bleak==0.20.2
bleak-winrt==1.2.0
blinker==1.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
classes==0.4.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.5
cryptography==39.0.0
cycler==0.11.0
decorator==4.4.2
Django==4.1.2
django-appconf==1.0.6
django-extensions==3.2.3
django-user-accounts==3.3.2
dm-tree==0.1.8
et-xmlfile==1.1.0
etils==1.2.0
exceptiongroup==1.1.3
ffmpeg==1.4
ffpyplayer==4.4.0
Flask==3.0.0
flatbuffers==23.1.21
fluidsynth==0.2
fonttools==4.37.4
forex-python==1.8
frozendict==2.3.4
frozenlist==1.3.3
future==0.18.2
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.51.1
h11==0.14.0
h5py==3.8.0
hidapi==0.13.1
html5lib==1.1
idna==3.4
imageio==2.25.1
imageio-ffmpeg==0.4.8
importlib-resources==5.12.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
keyboard==0.13.5
kiwisolver==1.4.4
lazy_loader==0.3
libclang==15.0.6.1
librosa==0.10.1
llvmlite==0.39.0
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.1
mido==1.3.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.2.1
mpy-cross-v5==1.0.1
mpy-cross-v6==1.0.1
msgpack==1.0.5
multidict==6.0.4
multitasking==0.0.11
mysql-connector-python==8.2.0
numba==0.56.0
numpy==1.22.4
oauthlib==3.2.2
openai==0.27.6
opencv-contrib-python==4.7.0.72
opencv-contrib-python-headless==4.7.0.72
opencv-python==4.8.0.76
opencv-python-headless==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==22.0
pandas==1.5.3
paramiko==3.3.1
pefile==2022.5.30
Pillow==9.2.0
platformdirs==3.10.0
playsound==1.3.0
pooch==1.7.0
proglog==0.1.10
promise==2.3
prompt-toolkit==3.0.41
protobuf==4.21.12
psutil==5.9.4
py2exe==0.13.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
PyAutoGUI==0.9.53
pybricks==3.3.0
pybricksdev==1.0.0a46
pycparser==2.21
pydub==0.25.1
pyFirmata==1.1.0
pygame==2.3.0
PyGetWindow==0.0.9
pyglet==2.0.4
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ev3dev2==2.1.0.post1
pytweening==1.0.4
pytz==2022.7.1
pyusb==1.2.1
pywin32-ctypes==0.2.0
reactivex==4.0.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
rtmidi==2.5.0
scikit-learn==1.3.0
scipy==1.8.1
selenium==4.11.2
semver==2.13.0
simplejson==3.19.1
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
Solution==5.4.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.6
sqlparse==0.4.3
sympy==1.10.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
tensorflow-metadata==1.13.1
termcolor==2.2.0
threadpoolctl==3.2.0
toml==0.10.2
torch==1.13.0
tqdm==4.64.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.4.0
tzdata==2022.4
urllib3==1.26.14
validators==0.18.2
wcwidth==0.2.12
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.6.0
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
yfinance==0.2.9
zipp==3.15.0
